flask-backend/requirements.txt
absl-py==2.0.0
appdirs==1.4.4
astunparse==1.6.3
attrs==23.1.0
blinker==1.6.3
cachetools==5.3.1
cattrs==23.1.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
exceptiongroup==1.1.3
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.43.1
gast==0.5.4
geojson-rewind==1.1.0
geomet==1.0.0
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.0
ml-dtypes==0.2.0
numpy==1.26.1
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
Pillow==10.1.0
platformdirs==3.11.0
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pygbif==0.6.3
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
requests-cache==1.1.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.3
six==1.16.0
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.14.0
termcolor==2.3.0
typing_extensions==4.8.0
url-normalize==1.4.3
urllib3==2.0.7
Werkzeug==3.0.0
wrapt==1.14.1
zipp==3.17.0
